deps: upgrade bisheng
