+++ package.json
@@ -74,7 +74,7 @@
     "babel-plugin-import": "^1.0.0",
     "babel-plugin-transform-runtime": "~6.15.0",
     "babel-preset-react-native": "^1.9.1",
-    "bisheng": "^0.19.0",
+    "bisheng": "^0.20.0",
     "bisheng-plugin-antd": "^0.8.0",
     "bisheng-plugin-description": "^0.1.1",
     "bisheng-plugin-react": "^0.4.0",
