+++ site/bisheng.common.config.js
@@ -2,24 +2,6 @@ const path = require('path');
 const configSvg = require('../svg.config');
 
 module.exports = {
-  lazyLoad(nodePath, nodeValue) {
-    if (typeof nodeValue === 'string') {
-      return true;
-    }
-    return nodePath.endsWith('/demo');
-  },
-  pick: {
-    components(markdownData) {
-      const filename = markdownData.meta.filename;
-      if (!/^components/.test(filename) ||
-          /\/demo$/.test(path.dirname(filename))) return;
-      /* eslint-disable consistent-return */
-      return {
-        meta: markdownData.meta,
-      };
-      /* eslint-enable consistent-return */
-    },
-  },
   webpackConfig(config) {
     configSvg(config, true);
     config.externals = {
